Refactor FormDisplay.tpl functions to static methods
